+++ coderd/database/migrations/000223_notifications_metrics.up.sql
@@ -1,2 +1,2 @@
 ALTER TABLE notification_messages
-ADD COLUMN queued_seconds INTEGER NULL;
+ADD COLUMN queued_seconds FLOAT NULL;
